requirements.txt
cartopy
cgen == 2020.1
clouddrift == 0.31.0
cmocean == 3.1.3
copernicusmarine == 1.0.1
hydra-zen == 0.12.1
jax == 0.4.25
jaxparrow == 0.1.3
numpy == 1.26.4
parcels == 3.0.2
psutil == 5.9.8
pymbolic == 2022.2
python-dotenv == 1.0.1
s3fs == 2024.2.0
toolz == 0.12.1
vaex == 4.17.0
xarray == 2024.1.1
